Fix INCBIN/BIN bug (unquoted filename with args)
asm6's documentation states that the following syntaxes are allowed:

INCBIN foo.bin
INCBIN foo.bin, $400
INCBIN foo.bin, $200, $2000

The 1st example works fine, but the 2nd and 3rd fail with a vague error
message: "Can't open file."  The filename attempted to be opened isn't
shown, so the user has no idea what's wrong (and had it been, I suspect
this bug would have been fixed a lot sooner).

However, if you put the filename part in double-quotes, all 3 forms work
correctly.

The bug: incbin() calls getfilename() (and is the only piece of code
that uses getfilename()).  getfilename() calls getword() with the 3rd
argument (mcheck) being 0.  An mcheck of 0 causes the function to *not*
call strtok() against "mathy letters", meaning it acts like strchr()
but with multiple letters that cause it to stop.  So in the case of
the string "foo.bin, $200, $2000", the string "foo.bin," (note trailing
comma!) is returned and treated as the filename.

The fix: extend getword() to support an mcheck of 2, which calls
strtok(dst,",") and thus only gets the filename portion (consider that
the size/offset arguments to INCBIN are comma-delimited).

I could have simply changed getword() to use an mcheck value of 1, but
that could potentially allow syntactically bad things to work, such as
"INCBIN foo.bin^ $200".

Next: as mentioned, filenames can be double-quoted, which allows for
use of filenames with spaces in them.  Document this fact.

Finally: add test_cases/incbin with some code and binary data files
to check all of the above.  There's no good way to implement these
tests (and I do not like "test frameworks" that bloat software even
more), so just refer to the code comments.